+++ linera-execution/src/unit_tests/applications_tests.rs
@@ -12,7 +12,7 @@ use linera_base::{
 };
 use serde::{Deserialize, Serialize};
 
-fn effect_id(index: u64) -> EffectId {
+fn effect_id(index: u32) -> EffectId {
     EffectId {
         chain_id: ChainId::root(0),
         height: BlockHeight::from(0),
@@ -20,18 +20,18 @@ fn effect_id(index: u64) -> EffectId {
     }
 }
 
-fn bytecode_id(index: u64) -> BytecodeId {
+fn bytecode_id(index: u32) -> BytecodeId {
     BytecodeId(effect_id(index))
 }
 
-fn app_id(index: u64) -> UserApplicationId {
+fn app_id(index: u32) -> UserApplicationId {
     UserApplicationId {
         bytecode_id: bytecode_id(0),
         creation: effect_id(index),
     }
 }
 
-fn app_description(index: u64, deps: Vec<u64>) -> UserApplicationDescription {
+fn app_description(index: u32, deps: Vec<u32>) -> UserApplicationDescription {
     UserApplicationDescription {
         bytecode_id: bytecode_id(0),
         bytecode_location: location(0),
@@ -41,7 +41,7 @@ fn app_description(index: u64, deps: Vec<u64>) -> UserApplicationDescription {
     }
 }
 
-fn location(operation_index: u64) -> BytecodeLocation {
+fn location(operation_index: u32) -> BytecodeLocation {
     #[derive(Serialize, Deserialize)]
     struct Dummy;
 
@@ -53,7 +53,7 @@ fn location(operation_index: u64) -> BytecodeLocation {
     }
 }
 
-fn registry(graph: impl IntoIterator<Item = (u64, Vec<u64>)>) -> ApplicationRegistry {
+fn registry(graph: impl IntoIterator<Item = (u32, Vec<u32>)>) -> ApplicationRegistry {
     let mut registry = ApplicationRegistry::default();
     for (index, deps) in graph {
         let description = app_description(index, deps);
